refactor with pool strategy
