+++ crates/burn-jit/src/kernel/pool/mod.rs
@@ -4,6 +4,8 @@ mod avg_pool2d_backward;
 mod base;
 mod max_pool2d;
 mod max_pool2d_backward;
+mod pool2d_shader;
+pub use pool2d_shader::*;
 
 pub(crate) use adaptive_avg_pool2d::*;
 pub use avg_pool2d::*;
