+++ crates/burn-jit/src/kernel/pool/pool2d_shader.rs
@@ -1,0 +1,238 @@
+use std::{marker::PhantomData, ops::Range};
+
+use crate::{
+    codegen::{Compilation, CompilationInfo, CompilationSettings, InputInfo, OutputInfo},
+    gpu::{gpu, Elem, Item, Scope, Variable, Visibility},
+    kernel::{DynamicKernelSource, SourceTemplate},
+    Compiler, JitElement, Runtime,
+};
+use std::fmt::Debug;
+
+pub(crate) trait PoolStrategy: Send + Sync + 'static + Clone + Debug {
+    type Accumulator: Copy;
+
+    fn h_range(&self) -> Range<u32>;
+    fn w_range(&self) -> Range<u32>;
+    fn initialize(&self, scope: &mut Scope, item: Item) -> Self::Accumulator;
+    fn process_result(
+        &self,
+        scope: &mut Scope,
+        accumulator: Self::Accumulator,
+        result: Variable,
+    ) -> Self::Accumulator;
+    fn assign(
+        &self,
+        scope: &mut Scope,
+        id: Variable,
+        output: Variable,
+        accumulator: Self::Accumulator,
+    );
+}
+
+pub(crate) struct Pool2dComputeShader<P: PoolStrategy, R: Runtime, E: JitElement> {
+    input: Variable,
+    output: Variable,
+    pool_strategy: P,
+    _elem: PhantomData<E>,
+    _runtime: PhantomData<R>,
+}
+
+impl<P: PoolStrategy, R: Runtime, E: JitElement> Pool2dComputeShader<P, R, E> {
+    fn expand(self, scope: &mut Scope) {
+        let input = self.input;
+        let output = self.output;
+        let id = Variable::Id;
+
+        let input_stride_0 = scope.create_local(Elem::UInt);
+        let input_stride_1 = scope.create_local(Elem::UInt);
+        let input_stride_2 = scope.create_local(Elem::UInt);
+        let input_stride_3 = scope.create_local(Elem::UInt);
+
+        let input_shape_0 = scope.create_local(Elem::UInt);
+        let input_shape_1 = scope.create_local(Elem::UInt);
+        let input_shape_2 = scope.create_local(Elem::UInt);
+        let input_shape_3 = scope.create_local(Elem::UInt);
+
+        let output_stride_0 = scope.create_local(Elem::UInt);
+        let output_stride_1 = scope.create_local(Elem::UInt);
+        let output_stride_2 = scope.create_local(Elem::UInt);
+        let output_stride_3 = scope.create_local(Elem::UInt);
+
+        let output_shape_0 = scope.create_local(Elem::UInt);
+        let output_shape_1 = scope.create_local(Elem::UInt);
+        let output_shape_2 = scope.create_local(Elem::UInt);
+        let output_shape_3 = scope.create_local(Elem::UInt);
+
+        gpu!(scope, input_stride_0 = stride(input, 0u32));
+        gpu!(scope, input_stride_1 = stride(input, 1u32));
+        gpu!(scope, input_stride_2 = stride(input, 2u32));
+        gpu!(scope, input_stride_3 = stride(input, 3u32));
+
+        gpu!(scope, input_shape_0 = shape(input, 2u32));
+        gpu!(scope, input_shape_1 = shape(input, 3u32));
+        gpu!(scope, input_shape_2 = shape(input, 2u32));
+        gpu!(scope, input_shape_3 = shape(input, 3u32));
+
+        gpu!(scope, output_stride_0 = stride(output, 0u32));
+        gpu!(scope, output_stride_1 = stride(output, 1u32));
+        gpu!(scope, output_stride_2 = stride(output, 2u32));
+        gpu!(scope, output_stride_3 = stride(output, 3u32));
+
+        gpu!(scope, output_shape_0 = shape(output, 0u32));
+        gpu!(scope, output_shape_1 = shape(output, 1u32));
+        gpu!(scope, output_shape_2 = shape(output, 2u32));
+        gpu!(scope, output_shape_3 = shape(output, 3u32));
+
+        let pool_stride_0 = Variable::GlobalScalar(0, Elem::UInt);
+        let pool_stride_1 = Variable::GlobalScalar(1, Elem::UInt);
+        let dilation_0 = Variable::GlobalScalar(2, Elem::UInt);
+        let dilation_1 = Variable::GlobalScalar(3, Elem::UInt);
+        let padding_0 = Variable::GlobalScalar(4, Elem::UInt);
+        let padding_1 = Variable::GlobalScalar(5, Elem::UInt);
+
+        let b = scope.create_local(Elem::UInt);
+        let c = scope.create_local(Elem::UInt);
+        let oh = scope.create_local(Elem::UInt);
+        let ow = scope.create_local(Elem::UInt);
+
+        gpu!(scope, b = id / output_stride_0);
+        gpu!(scope, b = b % output_shape_0);
+
+        gpu!(scope, c = id / output_stride_1);
+        gpu!(scope, c = c % output_shape_1);
+
+        gpu!(scope, oh = id / output_stride_2);
+        gpu!(scope, oh = oh % output_shape_2);
+
+        gpu!(scope, ow = id / output_stride_3);
+        gpu!(scope, ow = ow % output_shape_3);
+
+        let ih = scope.create_local(Elem::UInt);
+        let iw = scope.create_local(Elem::UInt);
+        let dilated = scope.create_local(Elem::UInt);
+
+        let ih_pad = scope.create_local(Elem::UInt);
+        let iw_pad = scope.create_local(Elem::UInt);
+        let result = scope.create_local(input.item());
+
+        let index_input = scope.create_local(Elem::UInt);
+        let index_input_0 = scope.create_local(Elem::UInt);
+        let index_input_1 = scope.create_local(Elem::UInt);
+        let index_input_2 = scope.create_local(Elem::UInt);
+        let index_input_3 = scope.create_local(Elem::UInt);
+
+        let within_padding_h = scope.create_local(Elem::Bool);
+        let within_padding_w = scope.create_local(Elem::Bool);
+        let tmp_padding = scope.create_local(Elem::Bool);
+        let border_bottom = scope.create_local(Elem::UInt);
+        let border_right = scope.create_local(Elem::UInt);
+
+        gpu!(scope, border_bottom = input_shape_2 + padding_0);
+        gpu!(scope, border_right = input_shape_3 + padding_1);
+
+        gpu!(scope, index_input_0 = b * input_stride_0);
+        gpu!(scope, index_input_1 = c * input_stride_1);
+
+        let accumulator = self.pool_strategy.initialize(scope, output.item());
+
+        self.pool_strategy.h_range().for_each(|kh| {
+            gpu!(scope, ih = oh * pool_stride_0);
+            gpu!(scope, dilated = kh * dilation_0);
+            gpu!(scope, ih += dilated);
+
+            gpu!(scope, within_padding_h = ih >= padding_0);
+            gpu!(scope, tmp_padding = ih < border_bottom);
+            gpu!(scope, within_padding_h = within_padding_h && tmp_padding);
+
+            gpu!(scope, if (within_padding_h).then(|scope| {
+                    self.pool_strategy.w_range().for_each(|kw| {
+                        gpu!(scope, iw = ow * pool_stride_1);
+                        gpu!(scope, dilated = kw * dilation_1);
+                        gpu!(scope, iw += dilated);
+
+                        gpu!(scope, within_padding_w = iw >= padding_1);
+                        gpu!(scope, tmp_padding = iw < border_right);
+                        gpu!(scope, within_padding_w = within_padding_w && tmp_padding);
+
+                        gpu!(scope, if (within_padding_w).then(|scope| {
+                            gpu!(scope, ih_pad = ih - padding_0);
+                            gpu!(scope, iw_pad = iw - padding_1);
+
+                            gpu!(scope, index_input_2 = ih_pad * input_stride_2);
+                            gpu!(scope, index_input_3 = iw_pad * input_stride_3);
+
+                            gpu!(scope, index_input = index_input_0);
+                            gpu!(scope, index_input += index_input_1);
+                            gpu!(scope, index_input += index_input_2);
+                            gpu!(scope, index_input += index_input_3);
+
+                            gpu!(scope, result = input[index_input]);
+
+                            self.pool_strategy.process_result(scope, accumulator, result);
+                        }));
+                    });
+            }));
+        });
+
+        self.pool_strategy.assign(scope, id, output, accumulator);
+    }
+}
+
+#[derive(new)]
+pub(crate) struct Pool2dEagerKernel<P: PoolStrategy, R: Runtime, E: JitElement> {
+    pool_strategy: P,
+    _runtime: PhantomData<R>,
+    _elem: PhantomData<E>,
+}
+
+impl<P: PoolStrategy, R: Runtime, E: JitElement> DynamicKernelSource
+    for Pool2dEagerKernel<P, R, E>
+{
+    fn source(&self) -> crate::kernel::SourceTemplate {
+        let mut scope = Scope::root();
+        let item = E::gpu_elem().into();
+
+        let input = Variable::GlobalInputArray(0, item);
+        let output = Variable::GlobalOutputArray(0, item);
+
+        scope.write_global_custom(output);
+
+        Pool2dComputeShader {
+            input,
+            output,
+            pool_strategy: self.pool_strategy.clone(),
+            _elem: PhantomData::<E>,
+            _runtime: PhantomData::<R>,
+        }
+        .expand(&mut scope);
+
+        let input = InputInfo::Array {
+            item,
+            visibility: Visibility::Read,
+        };
+        let scalars = InputInfo::Scalar {
+            elem: Elem::UInt,
+            size: 6,
+        };
+        let output = OutputInfo::Array { item };
+
+        let info = CompilationInfo {
+            inputs: vec![input, scalars],
+            outputs: vec![output],
+            scope,
+        };
+
+        let settings = CompilationSettings::default();
+        let shader = Compilation::new(info).compile(settings);
+        let shader = <R::Compiler as Compiler>::compile(shader);
+        SourceTemplate::new(shader.to_string())
+    }
+
+    fn id(&self) -> String {
+        format!(
+            "{:?}pool_strategy={:?}",
+            core::any::TypeId::of::<Self>(),
+            self.pool_strategy
+        )
+    }
+}
